requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.2
SQLAlchemy==2.0.12
WTForms==3.0.1
Werkzeug==2.3.4
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
greenlet==2.0.2
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
packaging==23.1
pip==23.1.2
pluggy==1.0.0
pytest==7.3.1
setuptools==67.7.2
typing-extensions==4.5.0
wheel==0.40.0
